ratelimits: aggressively rate-limit 'user-set-emailsent' action as well
This helps when an non-authnzerver email sending mechanism is being used
but still must be throttled. To handle this case, call the
'user-set-emailsent' authnzerver action before the email is
actually sent. This will return a 429 if too many requests
related to the specific email address in question have been made
within the last 60 seconds.